Correções de edição e persistencia de usuário com endereço correto

+++ desafioJava/src/main/java/com/desafioJava/controller/EnderecoController.java
@@ -37,7 +37,7 @@ public class EnderecoController {
 		enderecoFacade.excluir(endereco);
 	}
 
-	public Endereco buscarPorId(Long id) {
+	public Endereco buscarPorId(Integer id) {
 		return enderecoFacade.buscarPorId(id);
 	}
