+++ desafioJava/src/main/java/com/desafioJava/facade/EnderecoFacade.java
@@ -24,7 +24,7 @@ public class EnderecoFacade {
 		enderecoService.excluir(endereco);
 	}
 
-	public Endereco buscarPorId(Long id) {
+	public Endereco buscarPorId(Integer id) {
 		return enderecoService.buscarPorId(id);
 	}
